Feat/added deprecation warning to whole dataset drift check (#2011)
