+++ deepchecks/tabular/checks/__init__.py
@@ -20,9 +20,9 @@ from .model_evaluation import (BoostingOverfit, CalibrationScore, ConfusionMatri
                                TrainTestPerformance, TrainTestPredictionDrift, UnusedFeatures, WeakSegmentsPerformance)
 from .train_test_validation import (CategoryMismatchTrainTest, DatasetsSizeComparison, DateTrainTestLeakageDuplicates,
                                     DateTrainTestLeakageOverlap, FeatureLabelCorrelationChange,
-                                    IdentifierLabelCorrelation, IndexTrainTestLeakage, NewLabelTrainTest,
-                                    StringMismatchComparison, TrainTestFeatureDrift, TrainTestLabelDrift,
-                                    TrainTestSamplesMix, WholeDatasetDrift)
+                                    IdentifierLabelCorrelation, IndexTrainTestLeakage, MultivariateDrift,
+                                    NewLabelTrainTest, StringMismatchComparison, TrainTestFeatureDrift,
+                                    TrainTestLabelDrift, TrainTestSamplesMix, WholeDatasetDrift)
 
 __all__ = [
     # integrity checks
@@ -61,6 +61,7 @@ __all__ = [
     # distribution checks
     'TrainTestFeatureDrift',
     'TrainTestLabelDrift',
+    'MultivariateDrift',
     'WholeDatasetDrift',
     'TrainTestPredictionDrift',
