+++ deepchecks/tabular/deprecation_warnings.py
@@ -17,3 +17,10 @@ warnings.filterwarnings(
     category=DeprecationWarning,
     module=r'deepchecks.*'
 )
+
+warnings.filterwarnings(
+    action='once',
+    message=r'The WholeDatasetDrift check is deprecated.*',
+    category=DeprecationWarning,
+    module=r'deepchecks.*'
+)
