+++ docs/source/checks/tabular/train_test_validation/plot_train_test_feature_drift.py
@@ -31,7 +31,7 @@ How Deepchecks Detects Feature Drift
 This check detects feature drift by using :ref:`univariate measures <drift_detection_by_univariate_measure>`
 on each feature column separately.
 Another possible method for drift detection is by :ref:`a domain classifier <drift_detection_by_domain_classifier>`
-which is used in the :doc:`Whole Dataset Drift check </checks_gallery/tabular/train_test_validation/plot_whole_dataset_drift>`.
+which is used in the :doc:`Multivariate Drift check </checks_gallery/tabular/train_test_validation/plot_multivariate_drift>`.
 
 """
